# docs/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Radio Control Container API
  description: Minimal, stable contract for selecting a radio, setting channel and power, and receiving telemetry.
  version: 1.0.0
  contact:
    name: Radio Control Container Team
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/v1/capabilities:
    get:
      summary: Get system capabilities
      description: Returns the capabilities of the radio control system
      operationId: getCapabilities
      responses:
        '200':
          description: Capabilities retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                result: "ok"
                data:
                  telemetry: ["sse"]
                  commands: ["http-json"]
                  version: "1.0.0"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/radios:
    get:
      summary: List available radios
      description: Returns a list of all available radio devices
      operationId: listRadios
      responses:
        '200':
          description: Radios retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                result: "ok"
                data:
                  - id: "silvus-001"
                    name: "Silvus Radio 1"
                    status: "available"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/radios/select:
    post:
      summary: Select active radio
      description: Selects a radio as the active radio for operations
      operationId: selectRadio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - radioId
              properties:
                radioId:
                  type: string
                  description: ID of the radio to select
                  example: "silvus-001"
      responses:
        '200':
          description: Radio selected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Radio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/radios/{id}:
    get:
      summary: Get radio details
      description: Returns details for a specific radio
      operationId: getRadio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Radio ID
          example: "silvus-001"
      responses:
        '200':
          description: Radio details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid radio ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Radio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/telemetry:
    get:
      summary: Telemetry stream
      description: Server-Sent Events stream for real-time telemetry data
      operationId: getTelemetry
      responses:
        '200':
          description: Telemetry stream established
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE stream with telemetry events
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - result
        - data
      properties:
        result:
          type: string
          enum: ["ok"]
          description: Result status
        data:
          type: object
          description: Response data
        correlationId:
          type: string
          description: Correlation ID for request tracking
    ErrorResponse:
      type: object
      required:
        - result
        - code
        - message
      properties:
        result:
          type: string
          enum: ["error"]
          description: Result status
        code:
          type: string
          description: Error code
          enum: ["INVALID_RANGE", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "BUSY", "UNAVAILABLE", "INTERNAL"]
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
        correlationId:
          type: string
          description: Correlation ID for request tracking
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []